Move <CardIcon> to separate file
