+++ src/theme/DocCard/index.tsx
@@ -1,4 +1,4 @@
-import React, { type JSX } from 'react';
+import React from 'react';
 import clsx from 'clsx';
 import Link from '@docusaurus/Link';
 import { findFirstSidebarItemLink, useDocById } from '@docusaurus/theme-common/internal';
@@ -7,17 +7,8 @@ import { translate } from '@docusaurus/Translate';
 import type { Props } from '@theme/DocCard';
 import type { PropSidebarItemCategory, PropSidebarItemLink } from '@docusaurus/plugin-content-docs';
 import Heading from '@theme/Heading';
+import CardIcon from './CardIcon';
 import styles from './styles.module.css';
-import AmpIcon from '@site/static/img/amp.svg';
-import AndroidIcon from '@site/static/img/android.svg';
-import AngularIcon from '@site/static/img/angular.svg';
-import AppleIcon from '@site/static/img/apple.svg';
-import ChromecastIcon from '@site/static/img/chromecast.svg';
-import FlutterIcon from '@site/static/img/flutter.svg';
-import NuxtjsIcon from '@site/static/img/nuxtjs.svg';
-import ReactIcon from '@site/static/img/react.svg';
-import VuejsIcon from '@site/static/img/vuejs.svg';
-import WordPressIcon from '@site/static/img/wordpress.svg';
 
 function CardContainer({ href, children }) {
   return (
@@ -42,44 +33,17 @@ interface SidebarItemCustomProps {
   icon?: string;
 }
 
-interface CardIconProps {
+interface SidebarItemCardIconProps {
   item: PropSidebarItemCategory | PropSidebarItemLink;
   defaultIcon: string;
 }
 
-function CardIcon({ item, defaultIcon }: CardIconProps): JSX.Element | string | null {
-  const icon = (item.customProps as SidebarItemCustomProps)?.icon;
-  if (!icon) {
-    return defaultIcon;
-  }
-  switch (icon) {
-    case 'amp':
-      return <AmpIcon className={clsx(styles.cardIcon)} />;
-    case 'android':
-      return <AndroidIcon className={clsx(styles.cardIcon)} />;
-    case 'angular':
-      return <AngularIcon className={clsx(styles.cardIcon)} />;
-    case 'apple':
-      return <AppleIcon className={clsx(styles.cardIcon)} />;
-    case 'chromecast':
-      return <ChromecastIcon className={clsx(styles.cardIcon)} />;
-    case 'flutter':
-      return <FlutterIcon className={clsx(styles.cardIcon)} />;
-    case 'nuxtjs':
-      return <NuxtjsIcon className={clsx(styles.cardIcon)} />;
-    case 'react':
-      return <ReactIcon className={clsx(styles.cardIcon)} />;
-    case 'vuejs':
-      return <VuejsIcon className={clsx(styles.cardIcon)} />;
-    case 'wordpress':
-      return <WordPressIcon className={clsx(styles.cardIcon)} />;
-    default:
-      return icon;
-  }
+function SidebarItemCardIcon({ item, defaultIcon }: SidebarItemCardIconProps): JSX.Element {
+  return <CardIcon icon={(item.customProps as SidebarItemCustomProps)?.icon} defaultIcon={defaultIcon} />;
 }
 
 function CardCategory({ item }: { item: PropSidebarItemCategory }) {
-  const icon = <CardIcon item={item} defaultIcon="🗃️" />;
+  const icon = <SidebarItemCardIcon item={item} defaultIcon="🗃️" />;
   const href = findFirstSidebarItemLink(item);
   // Unexpected: categories that don't have a link have been filtered upfront
   if (!href) {
@@ -106,7 +70,7 @@ function CardCategory({ item }: { item: PropSidebarItemCategory }) {
 }
 
 function CardLink({ item }: { item: PropSidebarItemLink }) {
-  const icon = <CardIcon item={item} defaultIcon={isInternalUrl(item.href) ? '📄️' : '🔗'} />;
+  const icon = <SidebarItemCardIcon item={item} defaultIcon={isInternalUrl(item.href) ? '📄️' : '🔗'} />;
   const doc = useDocById(item.docId ?? undefined);
   return <CardLayout href={item.href} icon={icon} title={item.label} description={item.description ?? doc?.description} />;
 }
